% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.R
\name{eq}
\alias{eq}
\title{function that do equalization for audio.}
\usage{
eq(song, from, to, rate = 44100)
}
\arguments{
\item{song}{song file loaded}

\item{from}{frequency that you want to start}

\item{to}{frequency that you want to end}

\item{rate}{sampling rate of audio}
}
\value{
wave file with frequency you want to change
}
\description{
function that do equalization for audio.
}
\examples{
\dontrun{
ifonly <- loadSample("data/ifonly.wav")
ifonly_eq <- eq(
   song = ifonly,
   from = 400,
   to = 800,
   rate = 44100
 )
}

}
