% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melodyExtractor.R
\name{melodyExtractor}
\alias{melodyExtractor}
\title{function that remove the human vocal part melody}
\usage{
melodyExtractor(songLoadPath, songSavePath, compensate = FALSE,
  keyChange = NULL)
}
\arguments{
\item{songLoadPath}{path of the audio file}

\item{songSavePath}{path to save the song}

\item{compensate}{whether add low and high freq compensation}

\item{keyChange}{input semitones to change key}
}
\value{
A wav file without vocal part
}
\description{
function that remove the human vocal part melody
}
\examples{
\dontrun{
 melodyExtractor(
   songLoadPath = "inst/extdata/ifonly.wav",
   songSavePath = "novocal3.wav",
   compensate = FALSE,
   keyChange = +2
 )
}

}
