% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlePanEQ.R
\name{singlePanEQ}
\alias{singlePanEQ}
\title{function that do equalization for audio}
\usage{
singlePanEQ(sound, from, to, rate = 44100)
}
\arguments{
\item{sound}{single channel sound}

\item{from}{frequency that you want to start}

\item{to}{frequency that you want to end}

\item{rate}{sampling rate of audio}
}
\value{
single chanel sound file with frequency you want to change
}
\description{
function that do equalization for audio
}
\examples{
\dontrun{
 song <- sound::loadSample("inst/extdata/ifonly.wav")
 left_song <- sound::left(song)
 test <- singlePanEQ(
      left_song,
      from = 250,
      to = 8000,
      rate = 44100
 )
}

}
